requirements.txt
pandas==2.0.3
numpy==1.22.4
Pillow==9.1.1
tokenizers==0.13.3
scikit-learn==1.3.0
scipy==1.10.1
torch @ http://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp38-cp38-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu113/torchvision-0.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl
tornado==6.1
transformers==4.30.2
tqdm==4.65.0
